requirements.txt
PyQt6
pycryptodomex
pyserial
Pillow
qrcode
python-barcode
